(feature)add menu and merge maps
